requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==3.2.0; python_version >= '3.6'
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
flash==1.0.3
flask-bcrypt==0.7.1
flask==2.0.3
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
pycparser==2.21
pymysql==1.0.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==2.0.3; python_version >= '3.6'
